requirements.txt
streamlit
langchain
langchain-community
langchain-core
google-generativeai
pypdf
faiss-cpu
python-dotenv
